Create functions to move track to album and link track from album to artists

+++ sorter/sorter.py
@@ -11,12 +11,12 @@ class Sorter:
     """Sorter for GPM music.
 
     Attributes:
-        CATEGORIES (Tuple[str]): categories to sort music
-        metadata (Dict[str, Dict[str, str]]): metadata per track
+        metadata (Dict[str, Dict[str, str]]): metadata per track;
+            contains a subset from that of track_data when loaded
+        track_data (eyed3.id3.tag.Tag): track metadata; loaded from
+            find_track()
 
     """
-
-    CATEGORIES = ('Album', 'Artist')
 
     def __init__(self) -> None:
         """Initialize metadata, so that music files will only be read once."""
@@ -27,9 +27,16 @@ class Sorter:
         for file in config.TRACKS.glob('*.csv'):
             # Title,Album,Artist,Duration (ms),Rating,Play Count,Removed
             for self.row in TrackCSV(file).read():
-                self.find_track()
-                for category in self.CATEGORIES:
-                    self.make_dirs(category)
+                track = self.find_track()
+                if not track:
+                    continue
+
+                self.extract_images()
+                album = self.make_dirs('Album', self.row['Album'])
+                self.move_track(track, album)
+
+                self.make_dirs('Artist', self.row['Artist'])
+                self.link_track(track, album)
 
                 # Check whether the track features multiple artists;
                 # create the directories for each individual artist as well.
@@ -38,7 +45,7 @@ class Sorter:
                     for artist in artists:
                         self.make_dirs('Artist', artist)
 
-    def make_dirs(self, category: str, element: str) -> None:
+    def make_dirs(self, category: str, element: str) -> Path:
         """Make a directory given category and element.
 
         Because `Path.mkdir` is supplied the arguments `parents` and
@@ -48,9 +55,13 @@ class Sorter:
             category (str): 'Album' or 'Artist'
             element (str): e.g. an artist or an album
 
+        Returns:
+            Path: the directory created
+
         """
         out_dir = config.DEST / f'{category}s' / element
         out_dir.mkdir(parents=True, exist_ok=True)
+        return out_dir
 
     def find_track(self) -> Union[Path, None]:
         """Find the track given artist.
@@ -64,6 +75,8 @@ class Sorter:
         """
         artist = self.row['Artist']
 
+        # Before checking more new files, check whether the track
+        # has already been seen in self.metadata.
         for track, metadata in self.metadata.items():
             if (metadata['artist'] == artist
                     and metadata['album'] == self.row['Album']
@@ -95,13 +108,47 @@ class Sorter:
                     and self.track_data.tag.title == self.row['Title']):
                 return track
 
+        # Nothing was found?
+        del self.track_data
+        config.LOGGER.warning(
+            'Could not find a matching track file given the following:'
+            )
+        one_line = f'{self.row["Album"]}-{self.row["Title"]}-{artist}'
+        config.LOGGER.warning(one_line)
+
     def extract_images(self) -> None:
         """Extract album images from the track."""
         album = self.track_data.tag.album
         for image in self.track_data.tag.images:
             img_file = config.DEST / 'Albums' / album / image.makeFileName()
+            if img_file.exists():
+                continue
             with img_file.open(mode='wb') as img:
                 img.write(image.image_data)
+
+    def move_track(self, track: Path, album: Path) -> None:
+        """Move a track from the Takeout folder into its album.
+
+        Args:
+            track (Path): the file and path of the track
+            album (Path): the album directory
+
+        """
+        dest = album / track.name
+        track.replace(dest)
+
+    def link_track(self, track: Path, album: Path, artist: Path) -> None:
+        """Link a track from its album to artists' directories.
+
+        Args:
+            track (Path): the file and path of the track
+            album (Path): the album directory
+            artist (Path): the artist directory
+
+        """
+        src = album / track.name
+        dest = artist / track.name
+        dest.symlink_to(src)
 
 
 class TrackCSV:
